[Fix] 코드 충돌 해결 - #33
